Applying code review comments

+++ src/System.Diagnostics.Process/tests/ProcessTests.cs
@@ -145,7 +145,7 @@ namespace System.Diagnostics.Tests
             Win32Exception e = Assert.Throws<Win32Exception>(() => Process.Start(new ProcessStartInfo { UseShellExecute = true, FileName = fileToOpen }));
         }
 
-        [PlatformSpecific(TestPlatforms.Windows)] // Expected behavior varies on Windows and Unix. Refer to #23969
+        [PlatformSpecific(TestPlatforms.Windows)]
         [ConditionalTheory(typeof(PlatformDetection), nameof(PlatformDetection.HasWindowsShell)), InlineData(true), InlineData(false)]
         [SkipOnTargetFramework(TargetFrameworkMonikers.Uap, "not supported on UAP")]
         [OuterLoop("Launches File Explorer")]
@@ -170,10 +170,17 @@ namespace System.Diagnostics.Tests
                 }
                 else
                 {
-                    Assert.Equal("notepad", px.ProcessName);
+                    if (px != null)
+                    {
+                        Assert.Equal("notepad", px.ProcessName);
 
-                    px.Kill();
-                    Assert.True(px.WaitForExit(WaitInMS));
+                        px.Kill();
+                        Assert.True(px.WaitForExit(WaitInMS));
+                    }
+                    else
+                    {
+                        Console.WriteLine("Warning: Need to investigate why process is null when opening file on this machine. Refer to #24048");
+                    }
                 }
             }
         }
